test_requirements.txt
pytest==6.1.1
pytest-aiohttp==0.3.0
typing_extensions>=3.6.5